Move generic functionality to its own open graph image helper file
