+++ tests/doubles/presentations/indexable-presentation-double.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * WPSEO plugin test file.
  *
