+++ tests/presenters/opengraph/image-presenter-test.php
@@ -94,87 +94,6 @@ class Image_Presenter_Test extends TestCase {
 		);
 	}
 
-	/**
-	 * Tests the case where a 'non' relative url is given.
-	 *
-	 * @covers ::format_image
-	 */
-	public function test_format_image_with_string_given() {
-		$this->url_helper
-			->expects( 'is_relative' )
-			->once()
-			->with( 'image.jpg' )
-			->andReturnFalse();
-
-		$this->assertEquals(
-			[
-				'url' => 'image.jpg',
-			],
-			$this->instance->format_image( 'image.jpg' )
-		);
-	}
-
-	/**
-	 * Tests the case where a relative url is given.
-	 *
-	 * @covers ::format_image
-	 */
-	public function test_format_image_with_relative_image_url_given() {
-		$this->url_helper
-			->expects( 'is_relative' )
-			->once()
-			->with( 'image.jpg' )
-			->andReturnTrue();
-
-		$this->url_helper
-			->expects( 'get_relative_path' )
-			->once()
-			->with( 'image.jpg' )
-			->andReturn( 'https://example.org/image.jpg' );
-
-		$this->assertEquals(
-			[
-				'url' => 'https://example.org/image.jpg',
-			],
-			$this->instance->format_image( 'image.jpg' )
-		);
-	}
-
-	/**
-	 * Tests the case where the url was missing in the given param.
-	 *
-	 * @covers ::is_image_url_valid
-	 */
-	public function test_is_image_url_valid_with_missing_url() {
-		$this->assertFalse( $this->instance->is_image_url_valid( [] ) );
-	}
-
-	/**
-	 * Tests the case where the image is validated the right way (Happy path).
-	 *
-	 * @covers ::is_image_url_valid
-	 */
-	public function test_is_image_url_valid() {
-		$this->url_helper
-			->expects( 'get_extension_from_url' )
-			->once()
-			->with( 'image.jpg' )
-			->andReturn( 'jpg' );
-
-		$this->image_helper
-			->expects( 'is_extension_valid' )
-			->once()
-			->with( 'jpg' )
-			->andReturnTrue();
-
-		$this->assertTrue(
-			$this->instance->is_image_url_valid(
-				[
-					'url' => 'image.jpg',
-				]
-			)
-		);
-	}
 
 	/**
 	 * Tests the situation where the apply_filters returns a non-string value.
